+++ src/main/java/core/basesyntax/factory/UserDaoSingleFactory.java
@@ -1,7 +1,6 @@
 package core.basesyntax.factory;
 
 import core.basesyntax.dao.UserDao;
-import core.basesyntax.model.User;
 
 public class UserDaoSingleFactory {
     private static UserDao userDao;
