+++ src/main/java/core/basesyntax/lib/Injector.java
@@ -4,14 +4,13 @@ import core.basesyntax.dao.BetDao;
 import core.basesyntax.dao.UserDao;
 import core.basesyntax.factory.BetDaoSingleFactory;
 import core.basesyntax.factory.UserDaoSingleFactory;
-
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 
 public class Injector {
-    public static Object getInstance(Class clazz) throws NoSuchMethodException, IllegalAccessException,
-            InvocationTargetException, InstantiationException {
+    public static Object getInstance(Class clazz) throws NoSuchMethodException,
+            IllegalAccessException, InvocationTargetException, InstantiationException {
         Constructor constructor = clazz.getDeclaredConstructor();
         Object instance = constructor.newInstance();
